+++ app/src/main/java/com/codesoom/assignment/domain/ProductSaveRequest.java
@@ -1,9 +1,7 @@
-package com.codesoom.assignment.dto;
-
-import com.codesoom.assignment.domain.Product;
+package com.codesoom.assignment.domain;
 
 /**
- *  상품 등록에 필요한 데이터를 정의하는 인터페이스 입니다.
+ * 상품 등록에 필요한 데이터를 정의하는 인터페이스 입니다.
  */
 public interface ProductSaveRequest {
