+++ app/src/main/java/com/codesoom/assignment/dto/ProductSaveDto.java
@@ -1,5 +1,7 @@
 package com.codesoom.assignment.dto;
 
+import com.codesoom.assignment.domain.ProductSaveRequest;
+
 import java.beans.ConstructorProperties;
 
 /**
